+++ core/domain/stats_services_test.py
@@ -16,13 +16,9 @@
 
 import os
 
-from core.domain import event_services
-from core.domain import exp_domain
 from core.domain import exp_services
-from core.domain import stats_jobs_continuous
 from core.domain import stats_domain
 from core.domain import stats_services
-from core.domain import user_services
 from core.platform import models
 from core.tests import test_utils
 import feconf
@@ -43,9 +39,10 @@ class StatisticsServicesTest(test_utils.GenericTestBase):
             stats_models.ExplorationStatsModel.create('exp_id1', 1, 0, 0, {}))
 
     def test_calls_to_stats_methods(self):
-        """Test that calls are being made to the handle_stats_creation and
-        handle_new_exp_stats_creation methods when an exploration is created or
-        updated.
+        """Test that calls are being made to the
+        handle_stats_creation_for_new_exp_version and
+        handle_stats_creation_for_new_exploration methods when an exploration is
+        created or updated.
         """
         # Create exploration object in datastore.
         exp_id = 'exp_id'
@@ -88,8 +85,8 @@ class StatisticsServicesTest(test_utils.GenericTestBase):
             exploration_stats.state_stats_mapping.keys(), [
                 'Home', 'End', 'New state'])
 
-    def test_handle_new_exp_stats_creation(self):
-        """Test the handle_new_exp_stats_creation method."""
+    def test_handle_stats_creation_for_new_exploration(self):
+        """Test the handle_stats_creation_for_new_exploration method."""
         # Create exploration object in datastore.
         exp_id = 'exp_id'
         test_exp_filepath = os.path.join(
@@ -102,15 +99,15 @@ class StatisticsServicesTest(test_utils.GenericTestBase):
                 assets_list)
         exploration = exp_services.get_exploration_by_id(exp_id)
 
-        stats_services.handle_new_exp_stats_creation(exploration)
+        stats_services.handle_stats_creation_for_new_exploration(exploration)
 
         exploration_stats = stats_services.get_exploration_stats_by_id(
             exploration.id, exploration.version)
         self.assertEqual(
             exploration_stats.state_stats_mapping.keys(), ['Home', 'End'])
 
-    def test_handle_stats_creation(self):
-        """Test the handle_stats_creation method."""
+    def test_handle_stats_creation_for_new_exp_version(self):
+        """Test the handle_stats_creation_for_new_exp_version method."""
         # Create exploration object in datastore.
         exp_id = 'exp_id'
         test_exp_filepath = os.path.join(
@@ -133,7 +130,8 @@ class StatisticsServicesTest(test_utils.GenericTestBase):
             'cmd': 'add_state',
             'state_name': 'New state 2'
         }]
-        stats_services.handle_stats_creation(exploration, change_list)
+        stats_services.handle_stats_creation_for_new_exp_version(
+            exploration, change_list)
 
         exploration_stats = stats_services.get_exploration_stats_by_id(
             exploration.id, exploration.version)
@@ -155,7 +153,8 @@ class StatisticsServicesTest(test_utils.GenericTestBase):
             'old_state_name': 'New state 2',
             'new_state_name': 'Renamed state'
         }]
-        stats_services.handle_stats_creation(exploration, change_list)
+        stats_services.handle_stats_creation_for_new_exp_version(
+            exploration, change_list)
 
         exploration_stats = stats_services.get_exploration_stats_by_id(
             exploration.id, exploration.version)
@@ -170,7 +169,8 @@ class StatisticsServicesTest(test_utils.GenericTestBase):
             'cmd': 'delete_state',
             'state_name': 'New state'
         }]
-        stats_services.handle_stats_creation(exploration, change_list)
+        stats_services.handle_stats_creation_for_new_exp_version(
+            exploration, change_list)
 
         exploration_stats = stats_services.get_exploration_stats_by_id(
             exploration.id, exploration.version)
@@ -235,640 +235,3 @@ class StatisticsServicesTest(test_utils.GenericTestBase):
                     'total_solutions_triggered_count': 0
                 }
             })
-
-
-class ModifiedStatisticsAggregator(stats_jobs_continuous.StatisticsAggregator):
-    """A modified StatisticsAggregator that does not start a new batch
-    job when the previous one has finished.
-    """
-    @classmethod
-    def _get_batch_job_manager_class(cls):
-        return ModifiedStatisticsMRJobManager
-
-    @classmethod
-    def _kickoff_batch_job_after_previous_one_ends(cls):
-        pass
-
-
-class ModifiedStatisticsMRJobManager(
-        stats_jobs_continuous.StatisticsMRJobManager):
-
-    @classmethod
-    def _get_continuous_computation_class(cls):
-        return ModifiedStatisticsAggregator
-
-
-class ModifiedInteractionAnswerSummariesAggregator(
-        stats_jobs_continuous.StatisticsAggregator):
-    """A modified InteractionAnswerSummariesAggregator that does not start
-    a new batch job when the previous one has finished.
-    """
-    @classmethod
-    def _get_batch_job_manager_class(cls):
-        return ModifiedInteractionAnswerSummariesMRJobManager
-
-    @classmethod
-    def _kickoff_batch_job_after_previous_one_ends(cls):
-        pass
-
-
-class ModifiedInteractionAnswerSummariesMRJobManager(
-        stats_jobs_continuous.InteractionAnswerSummariesMRJobManager):
-
-    @classmethod
-    def _get_continuous_computation_class(cls):
-        return ModifiedInteractionAnswerSummariesAggregator
-
-
-class EventLogEntryTests(test_utils.GenericTestBase):
-    """Test for the event log creation."""
-
-    def test_create_events(self):
-        """Basic test that makes sure there are no exceptions thrown."""
-        event_services.StartExplorationEventHandler.record(
-            'eid', 2, 'state', 'session', {}, feconf.PLAY_TYPE_NORMAL)
-        event_services.MaybeLeaveExplorationEventHandler.record(
-            'eid', 2, 'state', 'session', 27.2, {}, feconf.PLAY_TYPE_NORMAL)
-
-
-class AnswerEventTests(test_utils.GenericTestBase):
-    """Test recording new answer operations through events."""
-
-    SESSION_ID = 'SESSION_ID'
-    TIME_SPENT = 5.0
-    PARAMS = {}
-
-    def test_record_answer(self):
-        self.save_new_default_exploration('eid', 'fake@user.com')
-        exp = exp_services.get_exploration_by_id('eid')
-
-        first_state_name = exp.init_state_name
-        second_state_name = 'State 2'
-        third_state_name = 'State 3'
-        exp_services.update_exploration('fake@user.com', 'eid', [{
-            'cmd': exp_domain.CMD_EDIT_STATE_PROPERTY,
-            'state_name': first_state_name,
-            'property_name': exp_domain.STATE_PROPERTY_INTERACTION_ID,
-            'new_value': 'TextInput',
-        }, {
-            'cmd': exp_domain.CMD_ADD_STATE,
-            'state_name': second_state_name,
-        }, {
-            'cmd': exp_domain.CMD_ADD_STATE,
-            'state_name': third_state_name,
-        }, {
-            'cmd': exp_domain.CMD_EDIT_STATE_PROPERTY,
-            'state_name': second_state_name,
-            'property_name': exp_domain.STATE_PROPERTY_INTERACTION_ID,
-            'new_value': 'TextInput',
-        }, {
-            'cmd': exp_domain.CMD_EDIT_STATE_PROPERTY,
-            'state_name': third_state_name,
-            'property_name': exp_domain.STATE_PROPERTY_INTERACTION_ID,
-            'new_value': 'Continue',
-        }], 'Add new state')
-        exp = exp_services.get_exploration_by_id('eid')
-
-        exp_version = exp.version
-
-        for state_name in [first_state_name, second_state_name]:
-            state_answers = stats_services.get_state_answers(
-                'eid', exp_version, state_name)
-            self.assertEqual(state_answers, None)
-
-        # answer is a string
-        event_services.AnswerSubmissionEventHandler.record(
-            'eid', exp_version, first_state_name, 'TextInput', 0, 0,
-            exp_domain.EXPLICIT_CLASSIFICATION, 'sid1', self.TIME_SPENT,
-            self.PARAMS, 'answer1')
-        event_services.AnswerSubmissionEventHandler.record(
-            'eid', exp_version, first_state_name, 'TextInput', 0, 1,
-            exp_domain.EXPLICIT_CLASSIFICATION, 'sid2', self.TIME_SPENT,
-            self.PARAMS, 'answer1')
-        # answer is a dict
-        event_services.AnswerSubmissionEventHandler.record(
-            'eid', exp_version, first_state_name, 'TextInput', 1, 0,
-            exp_domain.EXPLICIT_CLASSIFICATION, 'sid1', self.TIME_SPENT,
-            self.PARAMS, {'x': 1.0, 'y': 5.0})
-        # answer is a number
-        event_services.AnswerSubmissionEventHandler.record(
-            'eid', exp_version, first_state_name, 'TextInput', 2, 0,
-            exp_domain.EXPLICIT_CLASSIFICATION, 'sid1', self.TIME_SPENT,
-            self.PARAMS, 10)
-        # answer is a list of dicts
-        event_services.AnswerSubmissionEventHandler.record(
-            'eid', exp_version, first_state_name, 'TextInput', 3, 0,
-            exp_domain.EXPLICIT_CLASSIFICATION, 'sid1', self.TIME_SPENT,
-            self.PARAMS, [{'a': 'some', 'b': 'text'}, {'a': 1.0, 'c': 2.0}])
-        # answer is a list
-        event_services.AnswerSubmissionEventHandler.record(
-            'eid', exp_version, second_state_name, 'TextInput', 2, 0,
-            exp_domain.EXPLICIT_CLASSIFICATION, 'sid3', self.TIME_SPENT,
-            self.PARAMS, [2, 4, 8])
-        # answer is a unicode string
-        event_services.AnswerSubmissionEventHandler.record(
-            'eid', exp_version, second_state_name, 'TextInput', 1, 1,
-            exp_domain.EXPLICIT_CLASSIFICATION, 'sid4', self.TIME_SPENT,
-            self.PARAMS, self.UNICODE_TEST_STRING)
-        # answer is None (such as for Continue)
-        event_services.AnswerSubmissionEventHandler.record(
-            'eid', exp_version, third_state_name, 'Continue', 1, 1,
-            exp_domain.EXPLICIT_CLASSIFICATION, 'sid5', self.TIME_SPENT,
-            self.PARAMS, None)
-
-        expected_submitted_answer_list1 = [{
-            'answer': 'answer1', 'time_spent_in_sec': 5.0,
-            'answer_group_index': 0, 'rule_spec_index': 0,
-            'classification_categorization': 'explicit', 'session_id': 'sid1',
-            'interaction_id': 'TextInput', 'params': {}
-        }, {
-            'answer': 'answer1', 'time_spent_in_sec': 5.0,
-            'answer_group_index': 0, 'rule_spec_index': 1,
-            'classification_categorization': 'explicit', 'session_id': 'sid2',
-            'interaction_id': 'TextInput', 'params': {}
-        }, {
-            'answer': {'x': 1.0, 'y': 5.0}, 'time_spent_in_sec': 5.0,
-            'answer_group_index': 1, 'rule_spec_index': 0,
-            'classification_categorization': 'explicit', 'session_id': 'sid1',
-            'interaction_id': 'TextInput', 'params': {}
-        }, {
-            'answer': 10, 'time_spent_in_sec': 5.0, 'answer_group_index': 2,
-            'rule_spec_index': 0, 'classification_categorization': 'explicit',
-            'session_id': 'sid1', 'interaction_id': 'TextInput', 'params': {}
-        }, {
-            'answer': [{'a': 'some', 'b': 'text'}, {'a': 1.0, 'c': 2.0}],
-            'time_spent_in_sec': 5.0, 'answer_group_index': 3,
-            'rule_spec_index': 0, 'classification_categorization': 'explicit',
-            'session_id': 'sid1', 'interaction_id': 'TextInput', 'params': {}
-        }]
-        expected_submitted_answer_list2 = [{
-            'answer': [2, 4, 8], 'time_spent_in_sec': 5.0,
-            'answer_group_index': 2, 'rule_spec_index': 0,
-            'classification_categorization': 'explicit', 'session_id': 'sid3',
-            'interaction_id': 'TextInput', 'params': {}
-        }, {
-            'answer': self.UNICODE_TEST_STRING, 'time_spent_in_sec': 5.0,
-            'answer_group_index': 1, 'rule_spec_index': 1,
-            'classification_categorization': 'explicit', 'session_id': 'sid4',
-            'interaction_id': 'TextInput', 'params': {}
-        }]
-        expected_submitted_answer_list3 = [{
-            'answer': None, 'time_spent_in_sec': 5.0, 'answer_group_index': 1,
-            'rule_spec_index': 1, 'classification_categorization': 'explicit',
-            'session_id': 'sid5', 'interaction_id': 'Continue', 'params': {}
-        }]
-
-        state_answers = stats_services.get_state_answers(
-            'eid', exp_version, first_state_name)
-        self.assertEqual(
-            state_answers.get_submitted_answer_dict_list(),
-            expected_submitted_answer_list1)
-
-        state_answers = stats_services.get_state_answers(
-            'eid', exp_version, second_state_name)
-        self.assertEqual(
-            state_answers.get_submitted_answer_dict_list(),
-            expected_submitted_answer_list2)
-
-        state_answers = stats_services.get_state_answers(
-            'eid', exp_version, third_state_name)
-        self.assertEqual(
-            state_answers.get_submitted_answer_dict_list(),
-            expected_submitted_answer_list3)
-
-
-class RecordAnswerTests(test_utils.GenericTestBase):
-    """Tests for functionality related to recording and retrieving answers."""
-
-    EXP_ID = 'exp_id0'
-
-    def setUp(self):
-        super(RecordAnswerTests, self).setUp()
-        self.owner_id = self.get_user_id_from_email(self.OWNER_EMAIL)
-        user_services.create_new_user(self.owner_id, self.OWNER_EMAIL)
-        self.signup(self.OWNER_EMAIL, self.OWNER_USERNAME)
-        self.exploration = self.save_new_valid_exploration(
-            self.EXP_ID, self.owner_id, end_state_name='End')
-
-    def test_record_answer_without_retrieving_it_first(self):
-        stats_services.record_answer(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name, 'TextInput',
-            stats_domain.SubmittedAnswer(
-                'first answer', 'TextInput', 0,
-                0, exp_domain.EXPLICIT_CLASSIFICATION, {},
-                'a_session_id_val', 1.0))
-
-        state_answers = stats_services.get_state_answers(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name)
-        self.assertEqual(state_answers.get_submitted_answer_dict_list(), [{
-            'answer': 'first answer',
-            'time_spent_in_sec': 1.0,
-            'answer_group_index': 0,
-            'rule_spec_index': 0,
-            'classification_categorization': 'explicit',
-            'session_id': 'a_session_id_val',
-            'interaction_id': 'TextInput',
-            'params': {}
-        }])
-
-    def test_record_and_retrieve_single_answer(self):
-        state_answers = stats_services.get_state_answers(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name)
-        self.assertIsNone(state_answers)
-
-        stats_services.record_answer(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name, 'TextInput',
-            stats_domain.SubmittedAnswer(
-                'some text', 'TextInput', 0,
-                1, exp_domain.EXPLICIT_CLASSIFICATION, {},
-                'a_session_id_val', 10.0))
-
-        state_answers = stats_services.get_state_answers(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name)
-        self.assertEqual(state_answers.exploration_id, 'exp_id0')
-        self.assertEqual(state_answers.exploration_version, 1)
-        self.assertEqual(
-            state_answers.state_name, feconf.DEFAULT_INIT_STATE_NAME)
-        self.assertEqual(state_answers.interaction_id, 'TextInput')
-        self.assertEqual(state_answers.get_submitted_answer_dict_list(), [{
-            'answer': 'some text',
-            'time_spent_in_sec': 10.0,
-            'answer_group_index': 0,
-            'rule_spec_index': 1,
-            'classification_categorization': 'explicit',
-            'session_id': 'a_session_id_val',
-            'interaction_id': 'TextInput',
-            'params': {}
-        }])
-
-    def test_record_and_retrieve_single_answer_with_preexisting_entry(self):
-        stats_services.record_answer(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name, 'TextInput',
-            stats_domain.SubmittedAnswer(
-                'first answer', 'TextInput', 0,
-                0, exp_domain.EXPLICIT_CLASSIFICATION, {},
-                'a_session_id_val', 1.0))
-
-        state_answers = stats_services.get_state_answers(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name)
-        self.assertEqual(state_answers.get_submitted_answer_dict_list(), [{
-            'answer': 'first answer',
-            'time_spent_in_sec': 1.0,
-            'answer_group_index': 0,
-            'rule_spec_index': 0,
-            'classification_categorization': 'explicit',
-            'session_id': 'a_session_id_val',
-            'interaction_id': 'TextInput',
-            'params': {}
-        }])
-
-        stats_services.record_answer(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name, 'TextInput',
-            stats_domain.SubmittedAnswer(
-                'some text', 'TextInput', 0,
-                1, exp_domain.EXPLICIT_CLASSIFICATION, {},
-                'a_session_id_val', 10.0))
-
-        state_answers = stats_services.get_state_answers(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name)
-        self.assertEqual(state_answers.exploration_id, 'exp_id0')
-        self.assertEqual(state_answers.exploration_version, 1)
-        self.assertEqual(
-            state_answers.state_name, feconf.DEFAULT_INIT_STATE_NAME)
-        self.assertEqual(state_answers.interaction_id, 'TextInput')
-        self.assertEqual(state_answers.get_submitted_answer_dict_list(), [{
-            'answer': 'first answer',
-            'time_spent_in_sec': 1.0,
-            'answer_group_index': 0,
-            'rule_spec_index': 0,
-            'classification_categorization': 'explicit',
-            'session_id': 'a_session_id_val',
-            'interaction_id': 'TextInput',
-            'params': {}
-        }, {
-            'answer': 'some text',
-            'time_spent_in_sec': 10.0,
-            'answer_group_index': 0,
-            'rule_spec_index': 1,
-            'classification_categorization': 'explicit',
-            'session_id': 'a_session_id_val',
-            'interaction_id': 'TextInput',
-            'params': {}
-        }])
-
-    def test_record_many_answers(self):
-        state_answers = stats_services.get_state_answers(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name)
-        self.assertIsNone(state_answers)
-
-        submitted_answer_list = [
-            stats_domain.SubmittedAnswer(
-                'answer a', 'TextInput', 0, 1,
-                exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v', 10.0),
-            stats_domain.SubmittedAnswer(
-                'answer ccc', 'TextInput', 1, 1,
-                exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v', 3.0),
-            stats_domain.SubmittedAnswer(
-                'answer bbbbb', 'TextInput', 1, 0,
-                exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v', 7.5),
-        ]
-        stats_services.record_answers(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name, 'TextInput',
-            submitted_answer_list)
-
-        # The order of the answers returned depends on the size of the answers.
-        state_answers = stats_services.get_state_answers(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name)
-        self.assertEqual(state_answers.exploration_id, 'exp_id0')
-        self.assertEqual(state_answers.exploration_version, 1)
-        self.assertEqual(
-            state_answers.state_name, feconf.DEFAULT_INIT_STATE_NAME)
-        self.assertEqual(state_answers.interaction_id, 'TextInput')
-        self.assertEqual(state_answers.get_submitted_answer_dict_list(), [{
-            'answer': 'answer a',
-            'time_spent_in_sec': 10.0,
-            'answer_group_index': 0,
-            'rule_spec_index': 1,
-            'classification_categorization': 'explicit',
-            'session_id': 'session_id_v',
-            'interaction_id': 'TextInput',
-            'params': {}
-        }, {
-            'answer': 'answer ccc',
-            'time_spent_in_sec': 3.0,
-            'answer_group_index': 1,
-            'rule_spec_index': 1,
-            'classification_categorization': 'explicit',
-            'session_id': 'session_id_v',
-            'interaction_id': 'TextInput',
-            'params': {}
-        }, {
-            'answer': 'answer bbbbb',
-            'time_spent_in_sec': 7.5,
-            'answer_group_index': 1,
-            'rule_spec_index': 0,
-            'classification_categorization': 'explicit',
-            'session_id': 'session_id_v',
-            'interaction_id': 'TextInput',
-            'params': {}
-        }])
-
-    def test_record_answers_exceeding_one_shard(self):
-        # Use a smaller max answer list size so less answers are needed to
-        # exceed a shard.
-        with self.swap(
-            stats_models.StateAnswersModel, '_MAX_ANSWER_LIST_BYTE_SIZE',
-            100000):
-            state_answers = stats_services.get_state_answers(
-                self.EXP_ID, self.exploration.version,
-                self.exploration.init_state_name)
-            self.assertIsNone(state_answers)
-
-            submitted_answer_list = [
-                stats_domain.SubmittedAnswer(
-                    'answer a', 'TextInput', 0, 1,
-                    exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v',
-                    10.0),
-                stats_domain.SubmittedAnswer(
-                    'answer ccc', 'TextInput', 1, 1,
-                    exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v',
-                    3.0),
-                stats_domain.SubmittedAnswer(
-                    'answer bbbbb', 'TextInput', 1, 0,
-                    exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v',
-                    7.5),
-            ]
-            stats_services.record_answers(
-                self.EXP_ID, self.exploration.version,
-                self.exploration.init_state_name, 'TextInput',
-                submitted_answer_list * 200)
-
-            # Verify that more than 1 shard was stored. The index shard
-            # (shard_id 0) is not included in the shard count.
-            master_model = stats_models.StateAnswersModel.get_master_model(
-                self.exploration.id, self.exploration.version,
-                self.exploration.init_state_name)
-            self.assertGreater(master_model.shard_count, 0)
-
-            # The order of the answers returned depends on the size of the
-            # answers.
-            state_answers = stats_services.get_state_answers(
-                self.EXP_ID, self.exploration.version,
-                self.exploration.init_state_name)
-            self.assertEqual(state_answers.exploration_id, 'exp_id0')
-            self.assertEqual(state_answers.exploration_version, 1)
-            self.assertEqual(
-                state_answers.state_name, feconf.DEFAULT_INIT_STATE_NAME)
-            self.assertEqual(state_answers.interaction_id, 'TextInput')
-            self.assertEqual(
-                len(state_answers.get_submitted_answer_dict_list()), 600)
-
-    def test_record_many_answers_with_preexisting_entry(self):
-        stats_services.record_answer(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name, 'TextInput',
-            stats_domain.SubmittedAnswer(
-                '1 answer', 'TextInput', 0,
-                0, exp_domain.EXPLICIT_CLASSIFICATION, {},
-                'a_session_id_val', 1.0))
-
-        state_answers = stats_services.get_state_answers(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name)
-        self.assertEqual(state_answers.get_submitted_answer_dict_list(), [{
-            'answer': '1 answer',
-            'time_spent_in_sec': 1.0,
-            'answer_group_index': 0,
-            'rule_spec_index': 0,
-            'classification_categorization': 'explicit',
-            'session_id': 'a_session_id_val',
-            'interaction_id': 'TextInput',
-            'params': {}
-        }])
-
-        submitted_answer_list = [
-            stats_domain.SubmittedAnswer(
-                'answer aaa', 'TextInput', 0, 1,
-                exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v', 10.0),
-            stats_domain.SubmittedAnswer(
-                'answer ccccc', 'TextInput', 1, 1,
-                exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v', 3.0),
-            stats_domain.SubmittedAnswer(
-                'answer bbbbbbb', 'TextInput', 1, 0,
-                exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v', 7.5),
-        ]
-        stats_services.record_answers(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name, 'TextInput',
-            submitted_answer_list)
-
-        # The order of the answers returned depends on the size of the answers.
-        state_answers = stats_services.get_state_answers(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name)
-        self.assertEqual(state_answers.exploration_id, 'exp_id0')
-        self.assertEqual(state_answers.exploration_version, 1)
-        self.assertEqual(
-            state_answers.state_name, feconf.DEFAULT_INIT_STATE_NAME)
-        self.assertEqual(state_answers.interaction_id, 'TextInput')
-        self.assertEqual(state_answers.get_submitted_answer_dict_list(), [{
-            'answer': '1 answer',
-            'time_spent_in_sec': 1.0,
-            'answer_group_index': 0,
-            'rule_spec_index': 0,
-            'classification_categorization': 'explicit',
-            'session_id': 'a_session_id_val',
-            'interaction_id': 'TextInput',
-            'params': {}
-        }, {
-            'answer': 'answer aaa',
-            'time_spent_in_sec': 10.0,
-            'answer_group_index': 0,
-            'rule_spec_index': 1,
-            'classification_categorization': 'explicit',
-            'session_id': 'session_id_v',
-            'interaction_id': 'TextInput',
-            'params': {}
-        }, {
-            'answer': 'answer ccccc',
-            'time_spent_in_sec': 3.0,
-            'answer_group_index': 1,
-            'rule_spec_index': 1,
-            'classification_categorization': 'explicit',
-            'session_id': 'session_id_v',
-            'interaction_id': 'TextInput',
-            'params': {}
-        }, {
-            'answer': 'answer bbbbbbb',
-            'time_spent_in_sec': 7.5,
-            'answer_group_index': 1,
-            'rule_spec_index': 0,
-            'classification_categorization': 'explicit',
-            'session_id': 'session_id_v',
-            'interaction_id': 'TextInput',
-            'params': {}
-        }])
-
-
-class SampleAnswerTests(test_utils.GenericTestBase):
-    """Tests for functionality related to retrieving sample answers."""
-
-    EXP_ID = 'exp_id0'
-
-    def setUp(self):
-        super(SampleAnswerTests, self).setUp()
-        self.owner_id = self.get_user_id_from_email(self.OWNER_EMAIL)
-        user_services.create_new_user(self.owner_id, self.OWNER_EMAIL)
-        self.signup(self.OWNER_EMAIL, self.OWNER_USERNAME)
-        self.exploration = self.save_new_valid_exploration(
-            self.EXP_ID, self.owner_id, end_state_name='End')
-
-    def test_at_most_100_answers_returned_even_if_there_are_lots(self):
-        submitted_answer_list = [
-            stats_domain.SubmittedAnswer(
-                'answer a', 'TextInput', 0, 1,
-                exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v', 10.0),
-            stats_domain.SubmittedAnswer(
-                'answer ccc', 'TextInput', 1, 1,
-                exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v', 3.0),
-            stats_domain.SubmittedAnswer(
-                'answer bbbbb', 'TextInput', 1, 0,
-                exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v', 7.5),
-        ]
-        # Record 600 answers.
-        stats_services.record_answers(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name, 'TextInput',
-            submitted_answer_list * 200)
-
-        sample_answers = stats_services.get_sample_answers(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name)
-        self.assertEqual(len(sample_answers), 100)
-
-    def test_exactly_100_answers_returned_if_main_shard_has_100_answers(self):
-        submitted_answer_list = [
-            stats_domain.SubmittedAnswer(
-                'answer a', 'TextInput', 0, 1,
-                exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v', 10.0)
-        ]
-        # Record 100 answers.
-        stats_services.record_answers(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name, 'TextInput',
-            submitted_answer_list * 100)
-
-        sample_answers = stats_services.get_sample_answers(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name)
-        self.assertEqual(sample_answers, ['answer a'] * 100)
-
-    def test_all_answers_returned_if_main_shard_has_few_answers(self):
-        submitted_answer_list = [
-            stats_domain.SubmittedAnswer(
-                'answer a', 'TextInput', 0, 1,
-                exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v', 10.0),
-            stats_domain.SubmittedAnswer(
-                'answer bbbbb', 'TextInput', 1, 0,
-                exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v', 7.5),
-        ]
-        # Record 2 answers.
-        stats_services.record_answers(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name, 'TextInput',
-            submitted_answer_list)
-
-        sample_answers = stats_services.get_sample_answers(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name)
-        self.assertEqual(sample_answers, ['answer a', 'answer bbbbb'])
-
-    def test_only_sample_answers_in_main_shard_returned(self):
-        # Use a smaller max answer list size so fewer answers are needed to
-        # exceed a shard.
-        with self.swap(
-            stats_models.StateAnswersModel, '_MAX_ANSWER_LIST_BYTE_SIZE',
-            15000):
-            state_answers = stats_services.get_state_answers(
-                self.EXP_ID, self.exploration.version,
-                self.exploration.init_state_name)
-            self.assertIsNone(state_answers)
-
-            submitted_answer_list = [
-                stats_domain.SubmittedAnswer(
-                    'answer ccc', 'TextInput', 1, 1,
-                    exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v',
-                    3.0),
-            ]
-            stats_services.record_answers(
-                self.EXP_ID, self.exploration.version,
-                self.exploration.init_state_name, 'TextInput',
-                submitted_answer_list * 100)
-
-        # Verify more than 1 shard was stored. The index shard (shard_id 0)
-        # is not included in the shard count. Since a total of 100 answers were
-        # submitted, there must therefore be fewer than 100 answers in the
-        # index shard.
-        model = stats_models.StateAnswersModel.get('%s:%s:%s:%s' % (
-            self.exploration.id, str(self.exploration.version),
-            self.exploration.init_state_name, '0'))
-        self.assertEqual(model.shard_count, 1)
-
-        # Verify that the list of sample answers returned contains fewer than
-        # 100 answers, although a total of 100 answers were submitted.
-        sample_answers = stats_services.get_sample_answers(
-            self.EXP_ID, self.exploration.version,
-            self.exploration.init_state_name)
-        self.assertLess(len(sample_answers), 100)
